Organizing and placeholder project for libraries used by the webserver/ clients
