+++ .travis.yml
@@ -8,19 +8,19 @@ jobs:
       dist: bionic
       jdk: openjdk11
       script:
-        - ./gradlew BaseCode:check DesktopApp:check
+        - ./gradlew BaseCode:check DesktopApp:check WebServer:WebLibrary:check
         - ./gradlew codeCoverageReport
     - os: linux
       dist: bionic
       jdk: oraclejdk11
       script:
-        - ./gradlew BaseCode:check DesktopApp:check
+        - ./gradlew BaseCode:check DesktopApp:check WebServer:WebLibrary:check
         - ./gradlew codeCoverageReport
     - os: linux
       dist: xenial
       jdk: openjdk8
       script:
-        - ./gradlew BaseCode:check DesktopApp:check
+        - ./gradlew BaseCode:check DesktopApp:check WebServer:WebLibrary:check
         - ./gradlew codeCoverageReport
     # webserver tests with graal
     - os: linux
@@ -32,7 +32,7 @@ jobs:
       script:
         - source install-jdk.sh --url https://github.com/oracle/graal/releases/download/vm-19.2.1/graalvm-ce-linux-amd64-19.2.1.tar.gz # https://github.com/graalvm/graalvm-ce-builds/releases/download/vm-19.3.0/graalvm-ce-java11-linux-amd64-19.3.0.tar.gz
         - java -version
-        - ./gradlew WebServer:test
+        - ./gradlew WebServer:Server:test
         - ./gradlew codeCoverageReport
