--- WebServer/.dockerignore
@@ -1,4 +1,0 @@
-*
-!build/*-runner
-!build/*-runner.jar
-!build/lib/*
