+++ WebServer/Server/.dockerignore
@@ -1,0 +1,4 @@
+*
+!build/*-runner
+!build/*-runner.jar
+!build/lib/*
