+++ WebServer/Server/build.gradle
@@ -1,0 +1,58 @@
+// this block is necessary to make enforcedPlatform work for Quarkus plugin available
+// only locally (snapshot) that is also importing the Quarkus BOM
+buildscript {
+    repositories {
+        mavenLocal()
+    }
+    dependencies {
+        classpath "io.quarkus:quarkus-gradle-plugin:${quarkusPluginVersion}"
+    }
+}
+
+plugins {
+    id 'java'
+}
+
+apply plugin: 'io.quarkus'
+
+repositories {
+     mavenLocal()
+     mavenCentral()
+}
+
+dependencies {
+    compile project(":BaseCode:Core")
+    compile project(":WebServer:WebLibrary")
+
+    implementation 'io.quarkus:quarkus-mailer'
+    implementation 'io.quarkus:quarkus-resteasy-jackson'
+    implementation 'io.quarkus:quarkus-scheduler'
+    implementation 'io.quarkus:quarkus-elytron-security-properties-file'
+    implementation 'io.quarkus:quarkus-elytron-security'
+//    implementation 'io.quarkus:quarkus-mongodb-client'
+    implementation enforcedPlatform("${quarkusPlatformGroupId}:${quarkusPlatformArtifactId}:${quarkusPlatformVersion}")
+    implementation 'io.quarkus:quarkus-resteasy'
+
+    testImplementation 'io.quarkus:quarkus-junit5'
+    testImplementation 'io.rest-assured:rest-assured'
+
+    nativeTestImplementation 'io.quarkus:quarkus-junit5'
+    nativeTestImplementation 'io.rest-assured:rest-assured'
+}
+
+group 'com.gjs.taskTimekeeper.webServer'
+version = '1.0.0.2'
+
+compileJava {
+    options.compilerArgs << '-parameters'
+    println "WebServer version at compile time: " + version
+    dependsOn ':BaseCode:Core:test'
+    dependsOn ':WebServer:WebLibrary:test'
+    dependsOn 'incrementBuildVersion';
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
