+++ WebServer/Server/gradle.properties
@@ -1,0 +1,6 @@
+#Gradle properties
+#Wed Dec 04 00:19:31 UTC 2019
+quarkusPluginVersion=1.0.1.Final
+quarkusPlatformArtifactId=quarkus-universe-bom
+quarkusPlatformVersion=1.0.1.Final
+quarkusPlatformGroupId=io.quarkus
