+++ WebServer/Server/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+# Configuration file
+# key = value
