+++ WebServer/Server/src/native-test/java/com/gjs/taskTimekeeper/webServer/server/NativeExampleResourceIT.java
@@ -1,0 +1,9 @@
+package com.gjs.taskTimekeeper.webServer.server;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class NativeExampleResourceIT extends ExampleResourceTest {
+
+    // Execute the same tests but in native mode.
+}
