+++ WebServer/Server/src/test/java/com/gjs/taskTimekeeper/webServer/server/ExampleResourceTest.java
@@ -1,0 +1,21 @@
+package com.gjs.taskTimekeeper.webServer.server;
+
+import io.quarkus.test.junit.QuarkusTest;
+import org.junit.jupiter.api.Test;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.CoreMatchers.is;
+
+@QuarkusTest
+public class ExampleResourceTest {
+
+    @Test
+    public void testHelloEndpoint() {
+        given()
+          .when().get("/hello")
+          .then()
+             .statusCode(200)
+             .body(is("hello"));
+    }
+
+}
