+++ WebServer/WebLibrary/build.gradle
@@ -1,9 +1,10 @@
 plugins {
     id 'java'
     id "java-library"
+    id "io.freefair.lombok" version "4.1.5"
 }
 
-version = '1.0.1.50'
+version = '1.0.0.3'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
@@ -12,6 +13,7 @@ repositories {
 }
 
 dependencies {
+    compile project(":BaseCode:Core")
     //logging dependencies
     compile 'org.apache.logging.log4j:log4j-slf4j-impl:2.+'
     compile 'org.apache.logging.log4j:log4j-api:2.+'
@@ -22,12 +24,11 @@ dependencies {
     compile 'com.fasterxml.jackson.core:jackson-databind:2.9.+'
     compile 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.9.+'
 
-    compile 'args4j:args4j:2.33'
-
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
 
 compileJava {
-    println "Core version at compile time: " + version
+    println "WebLibrary at compile time: " + version
+    dependsOn ':BaseCode:Core:test'
     dependsOn 'incrementBuildVersion';
 }
