+++ WebServer/WebLibrary/src/main/java/com/gjs/taskTimekeeper/webServer/webLibrary/SomeClass.java
@@ -1,0 +1,7 @@
+package com.gjs.taskTimekeeper.webServer.webLibrary;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class SomeClass {
+}
