+++ WebServer/WebLibrary/src/test/resources/log4j2.yaml
@@ -1,0 +1,68 @@
+Configutation:
+  name: TestingLoggingConfiguration
+
+  Properties:
+    Property:
+      - name: archiveTriggeringSize
+        value: 1 KB
+      - name: archiveMaxFiles
+        value: 5
+      - name: logPath
+        value: "logs/test"
+      - name: logArchivePath
+        value: "/archive"
+      - name: logArchiveStampFormat
+        value: ".%d{yyyy-MM-dd-hh-mm-ss}.gz"
+      - name: logFileName
+        value: "/log.log"
+      - name: errorLogFileName
+        value: "/errorLog.log"
+      - name: logFile
+        value: "${logPath}${logFileName}"
+      - name: errorLogFile
+        value: "${logPath}${errorLogFileName}"
+      - name: logPattern
+        value: "%d %-5p [%t] - %m%n"
+      - name: logPatternDetailed
+        value: "%d %-5p [%t] (%-125l) - %m%n"
+      - name: logPatternColored
+        value: "%highlight{${logPattern}}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}"
+      - name: logPatternDetailedColored
+        value: "%highlight{${logPatternDetailed}}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}"
+  Appenders:
+    Console:
+      name: ConsoleOutput
+      target: SYSTEM_OUT
+      PatternLayout:
+        pattern: "${logPatternDetailedColored}"
+    RollingFile:
+      - name: LogFileOutput
+        fileName: ${logFile}
+        filePattern: "${logPath}${logArchivePath}${logFileName}${logArchiveStampFormat}"
+        PatternLayout:
+          pattern: "${logPatternDetailed}"
+        Policies:
+          SizeBasedTriggeringPolicy:
+            size: ${archiveTriggeringSize}
+        DefaultRollOverStrategy:
+          max: ${archiveMaxFiles}
+      - name: ErrorLogFileOutput
+        fileName: ${errorLogFile}
+        filePattern: "${logPath}${logArchivePath}${errorLogFileName}${logArchiveStampFormat}"
+        PatternLayout:
+          pattern: "${logPatternDetailed}"
+        Policies:
+          SizeBasedTriggeringPolicy:
+            size: ${archiveTriggeringSize}
+        DefaultRollOverStrategy:
+          max: ${archiveMaxFiles}
+  Loggers:
+    Root:
+      level: trace
+      AppenderRef:
+        - ref: ConsoleOutput
+          level: trace
+        - ref: LogFileOutput
+          level: trace
+        - ref: ErrorLogFileOutput
+          level: warn
