--- WebServer/build-run_notes.md
@@ -1,9 +1,0 @@
-# Task-Timekeeper Webserver Build/Run/Test
-
-## Prerequisites:
-
- - https://quarkus.io/guides/building-native-image-guide.html#prerequisites
- - Have Graalvm installed
-   - [Linux Instructions](https://gist.github.com/ricardozanini/fa65e485251913e1467837b1c5a8ed28)
- - Graalvm_home set
- - https://quarkus.io/guides/building-native-image-guide.html#prerequisites
