+++ WebServer/build.gradle
@@ -1,56 +1,1 @@
-// this block is necessary to make enforcedPlatform work for Quarkus plugin available
-// only locally (snapshot) that is also importing the Quarkus BOM
-buildscript {
-    repositories {
-        mavenLocal()
-    }
-    dependencies {
-        classpath "io.quarkus:quarkus-gradle-plugin:${quarkusPluginVersion}"
-    }
-}
-
-plugins {
-    id 'java'
-}
-
-apply plugin: 'io.quarkus'
-
-repositories {
-     mavenLocal()
-     mavenCentral()
-}
-
-dependencies {
-    implementation project(":BaseCode:ManagerIO")
-    implementation project(":BaseCode:Stats")
-
-    implementation 'io.quarkus:quarkus-smallrye-metrics'
-    implementation 'io.quarkus:quarkus-mailer'
-    implementation 'io.quarkus:quarkus-resteasy-jackson'
-    implementation 'io.quarkus:quarkus-scheduler'
-    implementation 'io.quarkus:quarkus-elytron-security'
-    implementation 'io.quarkus:quarkus-mongodb-client'
-    implementation enforcedPlatform("${quarkusPlatformGroupId}:${quarkusPlatformArtifactId}:${quarkusPlatformVersion}")
-    implementation 'io.quarkus:quarkus-resteasy'
-
-    testImplementation 'io.quarkus:quarkus-junit5'
-    testImplementation 'io.rest-assured:rest-assured'
-
-    nativeTestImplementation 'io.quarkus:quarkus-junit5'
-    nativeTestImplementation 'io.rest-assured:rest-assured'
-}
-
-group 'com.gjs.taskTimekeeper'
-version = '1.0.0.13'
-
-compileJava {
-    options.compilerArgs << '-parameters'
-    println "DesktopApp version at compile time: " + version
-    dependsOn ':BaseCode:test'
-    dependsOn 'incrementBuildVersion';
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
+build.dependsOn ":WebServer:WebLibrary:build", ":WebServer:Server:build"
