--- WebServer/gradle.properties
@@ -1,6 +1,0 @@
-#Gradle properties for WebServer
-# make sure matches in main settings.gradle plugin management
-quarkusPluginVersion=1.0.1.Final
-quarkusPlatformArtifactId=quarkus-universe-bom
-quarkusPlatformVersion=1.0.1.Final
-quarkusPlatformGroupId=io.quarkus
