--- WebServer/src/main/docker/Dockerfile.jvm
@@ -1,31 +1,0 @@
-####
-# This Dockerfile is used in order to build a container that runs the Quarkus application in JVM mode
-#
-# Before building the docker image run:
-#
-# mvn package
-#
-# Then, build the image with:
-#
-# docker build -f src/main/docker/Dockerfile.jvm -t quarkus/task-timekeeper-webserver-jvm .
-#
-# Then run the container using:
-#
-# docker run -i --rm -p 8080:8080 quarkus/task-timekeeper-webserver-jvm
-#
-###
-FROM fabric8/java-alpine-openjdk8-jre
-ENV JAVA_OPTIONS="-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager"
-ENV AB_ENABLED=jmx_exporter
-COPY build/lib/* /deployments/lib/
-COPY build/*-runner.jar /deployments/app.jar
-EXPOSE 8080
-
-# run with user 1001 and be prepared for be running in OpenShift too
-RUN adduser -G root --no-create-home --disabled-password 1001 \
-  && chown -R 1001 /deployments \
-  && chmod -R "g+rwX" /deployments \
-  && chown -R 1001:root /deployments
-USER 1001
-
-ENTRYPOINT [ "/deployments/run-java.sh" ]
