--- WebServer/src/main/docker/Dockerfile.native
@@ -1,22 +1,0 @@
-####
-# This Dockerfile is used in order to build a container that runs the Quarkus application in native (no JVM) mode
-#
-# Before building the docker image run:
-#
-# mvn package -Pnative -Dquarkus.native.container-build=true
-#
-# Then, build the image with:
-#
-# docker build -f src/main/docker/Dockerfile.native -t quarkus/task-timekeeper-webserver .
-#
-# Then run the container using:
-#
-# docker run -i --rm -p 8080:8080 quarkus/task-timekeeper-webserver
-#
-###
-FROM registry.access.redhat.com/ubi8/ubi-minimal
-WORKDIR /work/
-COPY build/*-runner /work/application
-RUN chmod 775 /work
-EXPOSE 8080
-CMD ["./application", "-Dquarkus.http.host=0.0.0.0"]
