--- WebServer/src/main/java/com/gjs/taskTimekeeper/webServer/ExampleResource.java
@@ -1,16 +1,0 @@
-package com.gjs.taskTimekeeper.webServer;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-@Path("/hello")
-public class ExampleResource {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "hello";
-    }
-}
