--- WebServer/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-# Configuration file
-# key = value
