--- WebServer/src/native-test/java/com/gjs/taskTimekeeper/webServer/NativeExampleResourceIT.java
@@ -1,9 +1,0 @@
-package com.gjs.taskTimekeeper.webServer;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeExampleResourceIT extends ExampleResourceTest {
-
-    // Execute the same tests but in native mode.
-}
