--- WebServer/src/test/java/com/gjs/taskTimekeeper/webServer/ExampleResourceTest.java
@@ -1,16 +1,0 @@
-package com.gjs.taskTimekeeper.webServer;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
-@QuarkusTest
-public class ExampleResourceTest {
-
-    @Test
-    public void testHelloEndpoint() {
-        given().when().get("/hello").then().statusCode(200).body(is("hello"));
-    }
-}
