+++ settings.gradle
@@ -32,5 +32,6 @@ include 'BaseCode:Core'
 include 'BaseCode:ManagerIO'
 include 'BaseCode:Stats'
 include 'DesktopApp'
-include 'WebServer'
+include 'WebServer:WebLibrary'
+include 'WebServer:Server'
 // include 'WebServer'
